Add support for the new brighter-fatter kernels
bfKernelNew is a temporary name for this dataset
while the crossover between old and new is handled
This is largely irrelevant though, as this code has
never fired before, as only obs_subaru actually
brighter-fatter correction, and it has its own isr
task. This will be updated once this dataset is
renamed.